Added OR conditions for future proofing
If we manage to make proper custom copspeech, then we can trim down some of these specific types to just the manufacturers. Is it a Lotus? Is it a Cobalt? Maybe a Pontiac? Or a Mini or Subaru? Manufacturer for some of these is not important; what is important is catching the crooks driving these... maybe.